+++ src/main/java/dev/rubric/discord/exceptions/UserExistsException.java
@@ -5,11 +5,11 @@ public class UserExistsException extends ApplicationException {
         super(String.format("User with username '%s' already exists", username), "USER_ALREADY_EXISTS");
     }
 
-    public UserExistsException(Integer userId) {
+    public UserExistsException(Long userId) {
         super(String.format("User with ID '%d' already exists", userId), "USER_ALREADY_EXISTS");
     }
 
-    public UserExistsException(Integer userId, Integer friendId) {
+    public UserExistsException(Long userId, Long friendId) {
         super(String.format("User '%d' is already friends with '%d'", userId, friendId), "FRIEND_ALREADY_EXISTS");
     }
 }
